Feat/transactions

+++ src/application/service/wallet/delete.ts
@@ -1,6 +1,7 @@
 import { Injectable } from "@nestjs/common";
 import { ApiProperty } from "@nestjs/swagger";
 import { PrismaUserWalletRepository, PrismaWalletRepository } from "src/infra/database/prisma-repositories/wallet";
+import { verifyIfUserIsWalletOwner } from "src/infra/utils/wallet/verifyIfUserIsOwner";
 
 export class DeleteWalletServiceDto {
   @ApiProperty({ type: "string", required: true })
@@ -15,7 +16,12 @@ export class DeleteWalletService {
   ) {}
 
   async execute({ walletId }: DeleteWalletServiceDto, userProfileId: string) {
-    await this.walletRepository.delete(walletId, userProfileId);
+    await verifyIfUserIsWalletOwner({
+      walletRepository: this.walletRepository,
+      walletId,
+      userProfileId
+    });
+    await this.walletRepository.delete(walletId);
     await this.userWalletRepository.delete(walletId, userProfileId);
   }
 }
